@@ -2,8 +2,8 @@
 {
     // This can easily be modified to be BaseEntity<T> and public T Id to support different key types.
     // Using non-generic integer types for simplicity and to ease caching logic
-    public class BaseEntity
+    public abstract class BaseEntity
     {
-        public int Id { get; set; }
+        public virtual int Id { get; protected set; }
     }
 }
