@@ -6,24 +6,34 @@ namespace Microsoft.eShopWeb.ApplicationCore.Entities.BasketAggregate
 {
     public class Basket : BaseEntity, IAggregateRoot
     {
-        public string BuyerId { get; set; }
+        public string BuyerId { get; private set; }
         private readonly List<BasketItem> _items = new List<BasketItem>();
         public IReadOnlyCollection<BasketItem> Items => _items.AsReadOnly();
+
+        public Basket(string buyerId)
+        {
+            BuyerId = buyerId;
+        }
 
         public void AddItem(int catalogItemId, decimal unitPrice, int quantity = 1)
         {
             if (!Items.Any(i => i.CatalogItemId == catalogItemId))
             {
-                _items.Add(new BasketItem()
-                {
-                    CatalogItemId = catalogItemId,
-                    Quantity = quantity,
-                    UnitPrice = unitPrice
-                });
+                _items.Add(new BasketItem(catalogItemId, quantity, unitPrice));
                 return;
             }
             var existingItem = Items.FirstOrDefault(i => i.CatalogItemId == catalogItemId);
-            existingItem.Quantity += quantity;
+            existingItem.AddQuantity(quantity);
+        }
+
+        public void RemoveEmptyItems()
+        {
+            _items.RemoveAll(i => i.Quantity == 0);
+        }
+
+        public void SetNewBuyerId(string buyerId)
+        {
+            BuyerId = buyerId;
         }
     }
 }
