@@ -2,8 +2,27 @@
 {
     public class BasketItem : BaseEntity
     {
-        public decimal UnitPrice { get; set; }
-        public int Quantity { get; set; }
-        public int CatalogItemId { get; set; }
+
+        public decimal UnitPrice { get; private set; }
+        public int Quantity { get; private set; }
+        public int CatalogItemId { get; private set; }
+        public int BasketId { get; private set; }
+
+        public BasketItem(int catalogItemId, int quantity, decimal unitPrice)
+        {
+            CatalogItemId = catalogItemId;
+            Quantity = quantity;
+            UnitPrice = unitPrice;
+        }
+
+        public void AddQuantity(int quantity)
+        {
+            Quantity += quantity;
+        }
+
+        public void SetNewQuantity(int quantity)
+        {
+            Quantity = quantity;
+        }
     }
 }
