@@ -1,11 +1,9 @@
-﻿using Microsoft.eShopWeb.ApplicationCore.Entities;
-
-namespace Microsoft.eShopWeb.ApplicationCore.Entities.BuyerAggregate
+﻿namespace Microsoft.eShopWeb.ApplicationCore.Entities.BuyerAggregate
 {
     public class PaymentMethod : BaseEntity
     {
-        public string Alias { get; set; }
-        public string CardId { get; set; } // actual card data must be stored in a PCI compliant system, like Stripe
-        public string Last4 { get; set; }
+        public string Alias { get; private set; }
+        public string CardId { get; private set; } // actual card data must be stored in a PCI compliant system, like Stripe
+        public string Last4 { get; private set; }
     }
 }
