@@ -1,7 +1,13 @@
-﻿namespace Microsoft.eShopWeb.ApplicationCore.Entities
+﻿using Microsoft.eShopWeb.ApplicationCore.Interfaces;
+
+namespace Microsoft.eShopWeb.ApplicationCore.Entities
 {
-    public class CatalogType : BaseEntity
+    public class CatalogType : BaseEntity, IAggregateRoot
     {
-        public string Type { get; set; }
+        public string Type { get; private set; }
+        public CatalogType(string type)
+        {
+            Type = type;
+        }
     }
 }
