@@ -1,6 +1,5 @@
 ﻿using Microsoft.eShopWeb.ApplicationCore.Interfaces;
 using Ardalis.GuardClauses;
-using Microsoft.eShopWeb.ApplicationCore.Entities;
 using System;
 using System.Collections.Generic;
 
@@ -34,11 +33,11 @@ namespace Microsoft.eShopWeb.ApplicationCore.Entities.OrderAggregate
         // but only through the method Order.AddOrderItem() which includes behavior.
         private readonly List<OrderItem> _orderItems = new List<OrderItem>();
 
-        public IReadOnlyCollection<OrderItem> OrderItems => _orderItems.AsReadOnly();
         // Using List<>.AsReadOnly() 
         // This will create a read only wrapper around the private list so is protected against "external updates".
         // It's much cheaper than .ToList() because it will not have to copy all items in a new collection. (Just one heap alloc for the wrapper instance)
         //https://msdn.microsoft.com/en-us/library/e78dcd75(v=vs.110).aspx 
+        public IReadOnlyCollection<OrderItem> OrderItems => _orderItems.AsReadOnly();
 
         public decimal Total()
         {
