@@ -1,6 +1,4 @@
-﻿using Microsoft.eShopWeb.ApplicationCore.Entities;
-
-namespace Microsoft.eShopWeb.ApplicationCore.Entities.OrderAggregate
+﻿namespace Microsoft.eShopWeb.ApplicationCore.Entities.OrderAggregate
 {
 
     public class OrderItem : BaseEntity
