@@ -4,13 +4,14 @@ using System.Threading.Tasks;
 
 namespace Microsoft.eShopWeb.ApplicationCore.Interfaces
 {
-    public interface IAsyncRepository<T> where T : BaseEntity
+    public interface IAsyncRepository<T> where T : BaseEntity, IAggregateRoot
     {
         Task<T> GetByIdAsync(int id);
-        Task<List<T>> ListAllAsync();
-        Task<List<T>> ListAsync(ISpecification<T> spec);
+        Task<IReadOnlyList<T>> ListAllAsync();
+        Task<IReadOnlyList<T>> ListAsync(ISpecification<T> spec);
         Task<T> AddAsync(T entity);
         Task UpdateAsync(T entity);
         Task DeleteAsync(T entity);
+        Task<int> CountAsync(ISpecification<T> spec);
     }
 }
