@@ -4,9 +4,8 @@ using System.Threading.Tasks;
 namespace Microsoft.eShopWeb.ApplicationCore.Interfaces
 {
 
-    public interface IOrderRepository : IRepository<Order>, IAsyncRepository<Order>
+    public interface IOrderRepository : IAsyncRepository<Order>
     {
-        Order GetByIdWithItems(int id);
         Task<Order> GetByIdWithItemsAsync(int id);
     }
 }
