@@ -1,7 +1,6 @@
 ﻿using Microsoft.eShopWeb.ApplicationCore.Entities.OrderAggregate;
 using Microsoft.eShopWeb.ApplicationCore.Interfaces;
 using Microsoft.EntityFrameworkCore;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace Microsoft.eShopWeb.Infrastructure.Data
@@ -12,20 +11,12 @@ namespace Microsoft.eShopWeb.Infrastructure.Data
         {
         }
 
-        public Order GetByIdWithItems(int id)
-        {
-            return _dbContext.Orders
-                .Include(o => o.OrderItems)
-                .Include($"{nameof(Order.OrderItems)}.{nameof(OrderItem.ItemOrdered)}")
-                .FirstOrDefault();
-        }
-
         public Task<Order> GetByIdWithItemsAsync(int id)
         {
             return _dbContext.Orders
                 .Include(o => o.OrderItems)
                 .Include($"{nameof(Order.OrderItems)}.{nameof(OrderItem.ItemOrdered)}")
-                .FirstOrDefaultAsync();
+                .FirstOrDefaultAsync(x => x.Id == id);
         }
     }
 }
