@@ -1,14 +1,23 @@
 ﻿using Microsoft.AspNetCore.Identity;
+using Microsoft.eShopWeb.ApplicationCore.Constants;
 using System.Threading.Tasks;
 
 namespace Microsoft.eShopWeb.Infrastructure.Identity
 {
     public class AppIdentityDbContextSeed
     {
-        public static async Task SeedAsync(UserManager<ApplicationUser> userManager)
+        public static async Task SeedAsync(UserManager<ApplicationUser> userManager, RoleManager<IdentityRole> roleManager)
         {
+            await roleManager.CreateAsync(new IdentityRole(AuthorizationConstants.Roles.ADMINISTRATORS));
+
             var defaultUser = new ApplicationUser { UserName = "demouser@microsoft.com", Email = "demouser@microsoft.com" };
-            await userManager.CreateAsync(defaultUser, "Pass@word1");
+            await userManager.CreateAsync(defaultUser, AuthorizationConstants.DEFAULT_PASSWORD);
+
+            string adminUserName = "admin@microsoft.com";
+            var adminUser = new ApplicationUser { UserName = adminUserName, Email = adminUserName };
+            await userManager.CreateAsync(adminUser, AuthorizationConstants.DEFAULT_PASSWORD);
+            adminUser = await userManager.FindByNameAsync(adminUserName);
+            await userManager.AddToRoleAsync(adminUser, AuthorizationConstants.Roles.ADMINISTRATORS);
         }
     }
 }
