@@ -1,11 +1,11 @@
-﻿using Microsoft.eShopWeb.ApplicationCore.Interfaces;
-using Microsoft.eShopWeb.Infrastructure.Identity;
-using Microsoft.AspNetCore.Authentication;
+﻿using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.eShopWeb.Web.ViewModels.Manage;
+using Microsoft.eShopWeb.ApplicationCore.Interfaces;
+using Microsoft.eShopWeb.Infrastructure.Identity;
 using Microsoft.eShopWeb.Web.Services;
+using Microsoft.eShopWeb.Web.ViewModels.Manage;
 using System;
 using System.Linq;
 using System.Text;
@@ -14,7 +14,8 @@ using System.Threading.Tasks;
 
 namespace Microsoft.eShopWeb.Web.Controllers
 {
-    [Authorize]
+    [ApiExplorerSettings(IgnoreApi = true)]
+    [Authorize] // Controllers that mainly require Authorization still use Controller/View; other pages use Pages
     [Route("[controller]/[action]")]
     public class ManageController : Controller
     {
@@ -24,7 +25,7 @@ namespace Microsoft.eShopWeb.Web.Controllers
         private readonly IAppLogger<ManageController> _logger;
         private readonly UrlEncoder _urlEncoder;
 
-        private const string AuthenicatorUriFormat = "otpauth://totp/{0}:{1}?secret={2}&issuer={0}&digits=6";
+        private const string AuthenticatorUriFormat = "otpauth://totp/{0}:{1}?secret={2}&issuer={0}&digits=6";
 
         public ManageController(
           UserManager<ApplicationUser> userManager,
@@ -44,7 +45,7 @@ namespace Microsoft.eShopWeb.Web.Controllers
         public string StatusMessage { get; set; }
 
         [HttpGet]
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> MyAccount()
         {
             var user = await _userManager.GetUserAsync(User);
             if (user == null)
@@ -490,7 +491,7 @@ namespace Microsoft.eShopWeb.Web.Controllers
         private string GenerateQrCodeUri(string email, string unformattedKey)
         {
             return string.Format(
-                AuthenicatorUriFormat,
+                AuthenticatorUriFormat,
                 _urlEncoder.Encode("eShopOnWeb"),
                 _urlEncoder.Encode(email),
                 unformattedKey);
