@@ -1,23 +1,12 @@
-﻿using Microsoft.eShopWeb.Web.Controllers;
-
-namespace Microsoft.AspNetCore.Mvc
+﻿namespace Microsoft.AspNetCore.Mvc
 {
     public static class UrlHelperExtensions
     {
         public static string EmailConfirmationLink(this IUrlHelper urlHelper, string userId, string code, string scheme)
         {
             return urlHelper.Action(
-                action: nameof(AccountController.ConfirmEmail),
-                controller: "Account",
-                values: new { userId, code },
-                protocol: scheme);
-        }
-
-        public static string ResetPasswordCallbackLink(this IUrlHelper urlHelper, string userId, string code, string scheme)
-        {
-            return urlHelper.Action(
-                action: nameof(AccountController.ResetPassword),
-                controller: "Account",
+                action: "GET",
+                controller: "ConfirmEmail",
                 values: new { userId, code },
                 protocol: scheme);
         }
