@@ -1,6 +1,4 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Microsoft.eShopWeb.Web.ViewModels.Manage
+﻿namespace Microsoft.eShopWeb.Web.ViewModels.Manage
 {
     public class TwoFactorAuthenticationViewModel
     {
