@@ -4,10 +4,7 @@
     ViewData["Title"] = "My Basket";
 }
 
-<a class="esh-basketstatus "
-   asp-area=""
-   asp-controller="Basket"
-   asp-action="Index">
+<a class="esh-basketstatus" asp-page="/Basket/Index">
         <div class="esh-basketstatus-image">
             <img src="~/images/cart.png" />
         </div>
